requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
openai-whisper
gradio
pytest
pytubefix